fix(input): correct Retina fullscreen clicks
Preserve fractional SDL window coordinates and map them through an aspect-fit viewport computed in logical window space. This avoids integer round-tripping through the physical Retina viewport and keeps production-button hit testing aligned in fullscreen.

Fixes #267

+++ GeneralsMD/Code/GameEngineDevice/Include/SDL3Device/GameClient/SDL3Mouse.h
@@ -84,7 +84,7 @@ private:
 
 	// Scale raw SDL window coordinates to game internal resolution
 	// GeneralsX @bugfix felipebraz 20/02/2026 Port fighter19 coordinate scaling fix
-	static void scaleMouseCoordinates(int rawX, int rawY, Uint32 windowID, int& scaledX, int& scaledY);
+	static void scaleMouseCoordinates(float rawX, float rawY, Uint32 windowID, int& scaledX, int& scaledY);
 
 	// Load cursor from ANI file (fighter19 pattern)
 	// GeneralsX @bugfix BenderAI 22/02/2026 Port fighter19 cursor loading
